requirements.txt
# Bibliothèques à installer
Flask ~= 3.0.0
customtkinter~=5.2.2
pillow~=10.3.0